Added new class InstrumentXLSResultsFileParser and refactored ChemStationParser to inherit from it
