+++ src/senaite/instruments/instruments/agilent/chemstation/instrument.py
@@ -1,30 +1,20 @@
 import csv
 import json
 import traceback
-import types
 from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims.exportimport.instruments import IInstrumentAutoImportInterface
 from bika.lims.exportimport.instruments import IInstrumentExportInterface
 from bika.lims.exportimport.instruments import IInstrumentImportInterface
+from bika.lims.exportimport.instruments.instrument import format_keyword
 from bika.lims.exportimport.instruments.resultsimport import AnalysisResultsImporter
-from bika.lims.exportimport.instruments.resultsimport import InstrumentCSVResultsFileParser
 from bika.lims.utils import t
 from cStringIO import StringIO
 from DateTime import DateTime
-from openpyxl import load_workbook
 from plone.i18n.normalizer.interfaces import IIDNormalizer
-from xlrd import open_workbook
+from senaite.instruments.instrument import InstrumentXLSResultsFileParser
 from zope.component import getUtility
 from zope.interface import implements
-from zope.publisher.browser import FileUpload
-
-
-class FileStub:
-    def __init__(self, file, name):
-        self.file = file
-        self.headers = {}
-        self.filename = name
 
 
 class chemstationexport(object):
@@ -100,83 +90,9 @@ class chemstationexport(object):
         request.RESPONSE.write(result)
 
 
-def xls_to_csv(infile, worksheet=0, delimiter=","):
-    # TODO: Move to utility module
-    """
-    Convert xlsx to easier format first, since we want to use the
-    convenience of the CSV library
-
-    """
-    def find_sheet(wb, worksheet):
-        for sheet in wb.sheets():
-            if sheet.name == worksheet:
-                return sheet
-
-    wb = open_workbook(file_contents=infile.read())
-    sheet = wb.sheets()[worksheet]
-
-    buffer = StringIO()
-
-    # extract all rows
-    for n, row in enumerate(sheet.get_rows()):
-        line = []
-        for cell in row:
-            value = cell.value
-            if type(value) in types.StringTypes:
-                value = value.encode("utf8")
-            if value is None:
-                value = ""
-            line.append(str(value))
-        print >>buffer, delimiter.join(line)
-    buffer.seek(0)
-    return buffer
-
-
-def xlsx_to_csv(infile, worksheet=0, delimiter=","):
-    # TODO: Move to utility module
-    """
-    Convert xlsx to easier format first, since we want to use the
-    convenience of the CSV library
-
-    """
-    wb = load_workbook(filename=infile)
-    sheet = wb.worksheets[worksheet]
-    buffer = StringIO()
-
-    # extract all rows
-    for n, row in enumerate(sheet.rows):
-        line = []
-        for cell in row:
-            value = cell.value
-            if type(value) in types.StringTypes:
-                value = value.encode("utf8")
-            if value is None:
-                value = ""
-            line.append(str(value))
-        print >>buffer, delimiter.join(line)
-    buffer.seek(0)
-    return buffer
-
-
-class ChemStationParser(InstrumentCSVResultsFileParser):
+class ChemStationParser(InstrumentXLSResultsFileParser):
     """ Parser
     """
-    def __init__(self, infile, fileformat='xlsx'):
-        InstrumentCSVResultsFileParser.__init__(self, infile, fileformat.upper())
-        # Convert xls to csv
-        self._delimiter = "|"
-        if fileformat == 'xlsx':
-            csv_data = xlsx_to_csv(
-                infile, worksheet=2, delimiter=self._delimiter)
-        elif fileformat == 'xls':
-            csv_data = xls_to_csv(
-                infile, worksheet=2, delimiter=self._delimiter)
-
-        # adpat csv_data into a FileUpload for parse method
-        stub = FileStub(file=csv_data, name=infile.filename)
-        self._infile = FileUpload(stub)
-        self._encoding = None
-        self._end_header = False
 
     def _parseline(self, line):
         if self._end_header:
@@ -229,22 +145,25 @@ class ChemStationParser(InstrumentCSVResultsFileParser):
         value_column = 'RT (min)'
         result = splitted[2]
         result = self.get_result(value_column, result, 0)
-        record[value_column] = result
+        record['ReturnTime'] = result
 
         value_column = 'Area'
         result = splitted[3]
         result = self.get_result(value_column, result, 0)
-        record[value_column] = result
+        record['Area'] = result
 
         value_column = 'Q-value'
         result = splitted[6]
         result = self.get_result(value_column, result, 0)
-        record[value_column] = result
+        record['QValue'] = result
 
         # assign record to kw dict
         kw = splitted[1]
+        kw = format_keyword(kw)
         ar_id = self._rawresults.keys()[0]
         self._rawresults[ar_id][0][kw] = record
+
+        return 0
 
     def get_result(self, column_name, result, line):
         result = str(result)
@@ -313,7 +232,7 @@ class chemstationimport(object):
         if not hasattr(infile, 'filename'):
             errors.append(_("No file selected"))
         if fileformat in ('xls', 'xlsx'):
-            parser = ChemStationParser(infile, fileformat=fileformat)
+            parser = ChemStationParser(infile, mimetype=fileformat)
         else:
             errors.append(t(_("Unrecognized file format ${fileformat}",
                               mapping={"fileformat": fileformat})))
